+++ src/mol-theme/color/carbohydrate-symbol.ts
@@ -1,23 +1,25 @@
 /**
- * Copyright (c) 2018-2020 mol* contributors, licensed under MIT, See LICENSE file for more info.
+ * Copyright (c) 2018-2026 mol* contributors, licensed under MIT, See LICENSE file for more info.
  *
  * @author Alexander Rose <alexander.rose@weirdbyte.de>
+ * @author Adam Midlik <midlik@gmail.com>
  */
 
-import { StructureElement, Bond, ElementIndex, Unit, Model } from '../../mol-model/structure';
-import { SaccharideColors, MonosaccharidesColorTable } from '../../mol-model/structure/structure/carbohydrates/constants';
 import { Location } from '../../mol-model/location';
-import type { ColorTheme, LocationColor } from '../color';
+import { Bond, ElementIndex, Model, StructureElement, Unit } from '../../mol-model/structure';
+import { MonosaccharidesColorTable, SaccharideColors, isSaccharideShapeDivided } from '../../mol-model/structure/structure/carbohydrates/constants';
 import { Color } from '../../mol-util/color';
-import { ParamDefinition as PD } from '../../mol-util/param-definition';
-import { ThemeDataContext } from '../theme';
 import { TableLegend } from '../../mol-util/legend';
+import { ParamDefinition as PD } from '../../mol-util/param-definition';
+import type { ColorTheme, LocationColor } from '../color';
+import { ThemeDataContext } from '../theme';
 import { ColorThemeCategory } from './categories';
+
 
 const DefaultColor = Color(0xCCCCCC);
 const Description = 'Assigns colors according to the Symbol Nomenclature for Glycans (SNFG).';
 
-export const CarbohydrateSymbolColorThemeParams = { };
+export const CarbohydrateSymbolColorThemeParams = {};
 export type CarbohydrateSymbolColorThemeParams = typeof CarbohydrateSymbolColorThemeParams
 export function getCarbohydrateSymbolColorThemeParams(ctx: ThemeDataContext) {
     return CarbohydrateSymbolColorThemeParams; // TODO return copy
@@ -29,21 +31,23 @@ export function CarbohydrateSymbolColorTheme(ctx: ThemeDataContext, props: PD.Va
     if (ctx.structure) {
         const { elements, getElementIndices } = ctx.structure.carbohydrates;
 
-        const getColor = (unit: Unit, index: ElementIndex) => {
+        const getColor = (unit: Unit, index: ElementIndex, isSecondary: boolean) => {
             if (!Unit.isAtomic(unit)) return DefaultColor;
             const carbs = getElementIndices(unit, index);
-            return carbs.length > 0 ? elements[carbs[0]].component.color : DefaultColor;
+            if (carbs.length === 0) return DefaultColor;
+            const component = elements[carbs[0]].component;
+            if (isSecondary && isSaccharideShapeDivided(component.type)) {
+                return SaccharideColors.Secondary;
+            } else {
+                return component.color;
+            }
         };
 
         color = (location: Location, isSecondary: boolean) => {
-            if (isSecondary) {
-                return SaccharideColors.Secondary;
-            } else {
-                if (StructureElement.Location.is(location)) {
-                    return getColor(location.unit, location.element);
-                } else if (Bond.isLocation(location)) {
-                    return getColor(location.aUnit, location.aUnit.elements[location.aIndex]);
-                }
+            if (StructureElement.Location.is(location)) {
+                return getColor(location.unit, location.element, isSecondary);
+            } else if (Bond.isLocation(location)) {
+                return getColor(location.aUnit, location.aUnit.elements[location.aIndex], isSecondary);
             }
             return DefaultColor;
         };